requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
fastapi==0.135.3
frozenlist==1.8.0
geoip2==5.2.0
h11==0.16.0
httptools==0.7.1
idna==3.11
markdown-it-py==4.0.0
maxminddb==3.1.1
mdurl==0.1.2
multidict==6.7.1
pip==26.0.1
propcache==0.4.1
psutil==7.2.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.1
rich==14.3.3
starlette==1.0.0
typing_extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.43.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
wsproto==1.3.2
yarl==1.23.0